requirements.txt
streamlit==1.29
google-generativeai
python-dotenv
nltk